clientAgent: entertainmentAuction value evaluation

+++ source/Tacy/src/ch/mas/tacy/model/auctions/ValuedAuction.java
@@ -6,9 +6,9 @@ import ch.mas.tacy.model.agentware.Auction;
 public class ValuedAuction implements Comparable<ValuedAuction> {
 
 	private final Auction auction;
-	private final double value;
+	private final float value;
 
-	public ValuedAuction(Auction auction, double value){
+	public ValuedAuction(Auction auction, float value){
 		this.auction = auction;
 		this.value = value;
 	}
@@ -19,7 +19,7 @@ public class ValuedAuction implements Comparable<ValuedAuction> {
 		return Double.compare(value, other.getValue());
 	}
 
-	public double getValue() {
+	public float getValue() {
 		return value;
 	}
